+++ src/melon/cli/commands/pxm/init_parser.py
@@ -1,14 +1,22 @@
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, override
 
-from ...base import BaseCommandProcessor
-from ...base.templates import BaseParameters
+from ....core.system_objects.manager.parser.enums import ExportStrategies
+from ...base.templates import T_SingleParserRequired
+from ._base import CommandProcessorTemplate
 
 if TYPE_CHECKING:
 	from dublib.cli.terminalyzer import CommandEntity, CommandModel
 
 	from ...base.structs import PreparedData
 
-class CommandProcessor(BaseCommandProcessor[BaseParameters]):
+@dataclass(frozen = True)
+class Parameters(T_SingleParserRequired):
+	"""Параметры, требуемые обработчиком."""
+
+	config_strategy: ExportStrategies
+
+class CommandProcessor(CommandProcessorTemplate[Parameters]):
 	"""Обработчик команды."""
 
 	#==========================================================================================#
@@ -26,6 +34,9 @@ class CommandProcessor(BaseCommandProcessor[BaseParameters]):
 		:rtype: CommandModel
 		"""
 
+		self._add_parser_position()
+		self._add_settings_export_strategy_position()
+
 		return model
 
 	@override
@@ -37,10 +48,10 @@ class CommandProcessor(BaseCommandProcessor[BaseParameters]):
 		:rtype: str
 		"""
 
-		return "Upgrade Melon."
+		return "Install parser."
 
 	@override
-	def _parse_parameters(self, entity: "CommandEntity", prepared_data: "PreparedData") -> BaseParameters:
+	def _parse_parameters(self, entity: "CommandEntity", prepared_data: "PreparedData") -> Parameters:
 		"""
 		Парсит данные обработанной команды в структуру **dataclass**.
 
@@ -49,23 +60,38 @@ class CommandProcessor(BaseCommandProcessor[BaseParameters]):
 		:param prepared_data: Подготовленные шаблонные параметры команды.
 		:type prepared_data: PreparedData
 		:return: Структура **dataclass**.
-		:rtype: BaseParameters
+		:rtype: Parameters
 		"""
 
-		return BaseParameters()
+		parameter = entity.get_position_parameter("STRATEGY", not_found_error = False)
+		strategy: str = "-s"
+
+		if hasattr(parameter, "name"):
+			strategy = getattr(parameter, "name")
+
+		return Parameters(
+			required_parser = prepared_data.required_parsers[0],
+			config_strategy = ExportStrategies(strategy)
+		)
 
 	@override
-	def _process(self, parameters: BaseParameters) -> bool:
+	def _process(self, parameters: Parameters) -> bool:
 		"""
 		Выполняет команду.
 
 		:param parameters: Параметры, требуемые обработчиком.
-		:type parameters: BaseParameters
+		:type parameters: Parameters
 		:return: Возвращает `True`, если выполнение успешно и прерывание не требуется.
 		:rtype: bool
 		"""
 
-		self.printer.emit(f"Repository: <i>{self.system_objects.options.REPOS.value}</i>")
-		self.system_objects.manager.upgrade()
+		repository_url: str = self.system_objects.manager.repositories.get(parameters.required_parser.name, exception = True)
+
+		self.printer.emit(f"Repository: <i>{repository_url}</i>.")
+		parameters.required_parser.install()
+		self.printer.emit("Parser installed.")
+
+		result = parameters.required_parser.export_settings(parameters.config_strategy)
+		self.printer.templates.manager.exported(result)
 
 		return True