+++ src/melon/utils/__init__.py
@@ -1,15 +1,13 @@
 from .cacher import Cacher
 from .classificator import Classificator
 from .collector import Collector
-from .development import DevelopmentAssistant
 from .timer import Timer
 from .unstubber import Unstubber
 
 __all__ = [
-    "Cacher",
-    "Classificator",
-    "Collector",
-    "DevelopmentAssistant",
-    "Timer",
-    "Unstubber"
+	"Cacher",
+	"Classificator",
+	"Collector",
+	"Timer",
+	"Unstubber"
 ]